backend/requirements.txt
asn1crypto==0.24.0
cffi==1.12.2
cryptography==2.6.1
Django==2.1.7
django-cors-headers==2.5.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
ecdsa==0.13
future==0.17.1
pyasn1==0.4.5
pycparser==2.19
PyJWT==1.7.1
python-jose==3.0.1
pytz==2018.9
rsa==4.0
six==1.12.0
